address comments from #226 (2)
