+++ lang/java/reef-io/src/test/java/org/apache/reef/io/network/group/impl/utils/ScatterHelperTest.java
@@ -18,21 +18,11 @@
  */
 package org.apache.reef.io.network.group.impl.utils;
 
-import org.apache.reef.io.network.group.api.task.CommunicationGroupServiceClient;
-import org.apache.reef.io.network.group.impl.driver.TopologySimpleNode;
-import org.apache.reef.io.network.util.StringIdentifierFactory;
-import org.apache.reef.io.serialization.Codec;
-import org.apache.reef.io.serialization.SerializableCodec;
-import org.apache.reef.wake.Identifier;
-import org.apache.reef.wake.IdentifierFactory;
 import org.junit.Test;
 
-import java.util.LinkedList;
 import java.util.List;
-import java.util.Map;
 
 import static org.junit.Assert.*;
-import static org.mockito.Mockito.*;
 
 /**
  * Tests for util classes related to Scatter.
@@ -62,57 +52,5 @@ public final class ScatterHelperTest {
       assertEquals(elementCount, sum); // all elements were considered check
       assertEquals(maxVal - minVal, elementCount % taskCount == 0 ? 0 : 1); // uniform distribution check
     }
-  }
-
-  /**
-   * Test that {@code ScatterEncoder} and {@code ScatterDecoder} function correctly.
-   * Create a small topology of 4 nodes and simulate a scatter operation.
-   */
-  @Test
-  public void testEncodeDecode() {
-    final IdentifierFactory ifac = new StringIdentifierFactory();
-    final Codec<Integer> codec = new SerializableCodec<>();
-
-    final List<Integer> elements = new LinkedList<>();
-    for (int element = 0; element < 400; element++) {
-      elements.add(element);
-    }
-
-    final List<Integer> counts = new LinkedList<>();
-    final List<Identifier> taskOrder = new LinkedList<>();
-    for (int index = 0; index < 4; index++) {
-      counts.add(100);
-      taskOrder.add(ifac.getNewInstance("Task-" + index));
-    }
-
-    final TopologySimpleNode rootNode = new TopologySimpleNode("Task-0");
-    final TopologySimpleNode childNode1 = new TopologySimpleNode("Task-1");
-    final TopologySimpleNode childNode2 = new TopologySimpleNode("Task-2");
-    final TopologySimpleNode childNode3 = new TopologySimpleNode("Task-3");
-    rootNode.addChild(childNode1);
-    rootNode.addChild(childNode2);
-    childNode1.addChild(childNode3);
-
-    final CommunicationGroupServiceClient mockCommGroupClient = mock(CommunicationGroupServiceClient.class);
-    when(mockCommGroupClient.getTopologySimpleNodeRoot()).thenReturn(rootNode);
-    final ScatterEncoder scatterEncoder = new ScatterEncoder(mockCommGroupClient);
-    final ScatterDecoder scatterDecoder = new ScatterDecoder();
-
-    final Map<String, byte[]> encodedDataMap = scatterEncoder.encode(elements, counts, taskOrder, codec);
-
-    // check msg correctness for childNode1 (Task-1)
-    ScatterData scatterData = scatterDecoder.decode(encodedDataMap.get(childNode1.getTaskId()));
-    for (int index = 0; index < 100; index++) {
-      assertTrue(index + 100 == codec.decode(scatterData.getMyData()[index]));
-    }
-    assertTrue(scatterData.getChildrenData().containsKey("Task-3"));
-    assertEquals(scatterData.getChildrenData().size(), 1);
-
-    // check msg correctness for childNode2 (Task-2)
-    scatterData = scatterDecoder.decode(encodedDataMap.get(childNode2.getTaskId()));
-    for (int index = 0; index < 100; index++) {
-      assertTrue(index + 200 == codec.decode(scatterData.getMyData()[index]));
-    }
-    assertTrue(scatterData.getChildrenData().isEmpty());
   }
 }
